feat: New Settings page
